Convversation_Memory/requirements.txt
streamlit 
googel-generativeai
python-dotenv
langchian
PyPDF2
faiss-cpu
langchian_google_genai